Deal with passive time lower of the hardware limit
Within the used configuration we need low time of at least 1e-7 s
and max 53.6870911875 s. We have not experienced this problem yet
because the trigger/gate was always used with experimental channels
reporting latency time higher than 1e-7 s. But, for example,
the sardana-limaccd controller is not reporting any latency time,
hence sardana uses value 0. Use the minimum passive time when repeat = 1,
e.g. ct - here it is not relevant, just an extra dead time,
and raise an exception if repat > 1.